refactor: split dc-api session apart for maintainability
